Unnecessary cast from int to long.
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/compress/trunk@1534269 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/compress/archivers/tar/TarArchiveInputStream.java
@@ -581,7 +581,7 @@ public class TarArchiveInputStream extends ArchiveInputStream {
         if (totalRead == -1) {
             hasHitEOF = true;
         } else {
-            entryOffset += (long) totalRead;
+            entryOffset += totalRead;
         }
 
         return totalRead;
